GPS_requirements.txt
geographiclib==2.1
geopy==2.4.1
pynmea2==1.19.0
pyserial==3.5
pytz==2025.2
